// src/app/providers/interfaces/theme.interface.ts
export type Theme =
  | 'default'
  | 'neon'
  | 'cyberpunk'
  | 'mystic'
  | 'burnt-forest'
  | 'paper'
  | 'paper-mario'
  | 'retro-80s'
  | 'pixel'
  | 'matrix'
  | 'terminal'
  | 'vaporwave'
  | 'arcade'
  | 'valentines'
  | 'st-patricks'
  | 'earth-day'
  | 'independence'
  | 'halloween'
  | 'thanksgiving'
  | 'christmas'
  | string;